refactor(planner): decompose WeeklyPlanner into façade + 4 mixins
Apply the established façade + mixins pattern to weekly_planner.py (1336 lines),
extracting 10 methods into 4 mixins under workflows/planner/:
- ContextLoadingMixin: file I/O for references, bilan, workout analyses
- PeriodizationMixin: training phase, PID, mesocycle, Zwift catalog
- PromptMixin: full planning prompt assembly
- OutputMixin: clipboard, session status, JSON persistence

Façade retains __init__, API init, week utilities, metrics, run(), and main().
No behavioral change — class name and all external imports unchanged.
