+++ magma_cycling/workflows/planner/__init__.py
@@ -1,0 +1,1 @@
+"""WeeklyPlanner mixin modules for god class decomposition."""
